+++ src/index.ts
@@ -2,3 +2,4 @@ export * from './glassnode-api.js';
 export * from './errors.js';
 export * from './types/config.js';
 export * from './types/metadata.js';
+export * from './types/params.js';